gfx/cro.c: only use Pango if there's a ~ in the string
Since the main reason for switching to Pango is that we'll need its
layout analysis capabilities to draw overlines, how about just not
using Pango for strings that don't have such complications ?

Time to first index is now 487 ms, almost as good as with Pango
disabled.

Note: a further optimization would be to check strings for special
issues at the source, and avoid all the strchr-ing.