file/git-hist.c (n_branches, find_commit): get rid of "open branches" array
We used to have an arrary of branches we've entered but not left yet,
which we used to look for commits already in the tree we're in the
process of building.

However, this didn't work as intended since a branch may have
progressed already past the current position and we wouldn't see the
commits on it. Furthermore, recursing the tree above us for each
branch could easily be less efficient than just running through all
the nodes.

So keeping a list of nodes is what we do now.