file/git-hist.c, gui/: support multiple heads; update GUI accordingly
While we try to iterate HEAD first, this won't result in HEAD coming
out on top if there is another branch that is ahead of HEAD. While
we could try some heuristics to avoid this, it'll be better to just
sort the whole list first by (partial) git history order and then by
date. That way, the result will never be too far from what a human
user will expect to see.