file/git-hist.c (vcs_git_history): limit depth of history retrieval
We used to retrieve the entire history, but if we afterwards only use
the topmost N entries, we need - in the worse case - only the first N
entries from each branch.

We could try to be even more efficient, but that would require doing
part of the sorting already during retrieval, which would make things
messy.