gui/comp.c: restructure information flow; pop-up now disappears on click
The goal is to provide visual feedback when clicking on a pop-up
item, since starting a browser or loading a large PDF can take
long enough for the user to become uncertain whether the click
was actually processed.

For this, we need to pass "gui" to comp_click, which is complicated
by there only being one user-provided argument, which we already used
for the URL.

We now have a per-item structure (which is passed to comp_click) that
provides both URL and "gui" pointer, and also makes a few other things
easier.