file/git-file.c (vcs_git_time): getting the right date is a little harder ...
First, we should be using the author ("content") date, not the commit
("envelope") date:
http://stackoverflow.com/questions/11856983/why-git-authordate-is-different-from-commitdate

Second, the date of the commit where we picked up the file doesn't
seem to be very relevant. What we're really interested in is when
the file last changed. For this, we need to perform an expensive
search.