file/: implement caching version of git_repository_open_ext for libgit2 < 0.22
In version 0.22, this was added:
Share open packfiles across repositories to share descriptors and mmaps

Unfortunately, some people are unable to upgrade to modern libraries.
Without FD sharing, eeshow may run out of file descriptors when
loading large repositories.

This commit adds a cache to reduce the number of times a repository
is opened, which may in turn reduce the number of open files.